Feature/single layer unittests (#366)
Do not link against the whole cio library in unit tests.

+++ src/linux/cio_linux_random.c
@@ -44,7 +44,7 @@ enum cio_error cio_entropy_get_bytes(void *bytes, size_t num_bytes)
 		return (enum cio_error)(-errno);
 	}
 
-	size_t ret = fread_unlocked(bytes, 1, num_bytes, dev_urandom);
+	size_t ret = fread_unlocked(bytes, (size_t)1, num_bytes, dev_urandom);
 	enum cio_error err;
 	if (cio_unlikely(ret < num_bytes)) {
 		err = CIO_EOF;
